# web endpoint optimization: or, the unreasonable effectiveness of asyncio.gather?