@thundergolfer 2024-02-18
did I `await` that gather?

<img width="318" alt="image" src="https://github.com/modal-labs/modal-client/assets/12058921/3d6a137d-3086-465e-8d07-80c61527ee91">